int to long

+++ src/main/java/com/jiffydelivery/jiffy/Entity/DBDAO/Customer.java
@@ -6,7 +6,6 @@ import java.util.List;
 
 @Entity
 @Table(name = "Customer")
-
 public class Customer implements Serializable {
 
     private static final long serialVersionUID = 7596499365251992523L;
@@ -42,11 +41,11 @@ public class Customer implements Serializable {
         return serialVersionUID;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
